Fix POM and migrate to JUnit 5
