+++ src/test/java/automation/CucumberTest.java
@@ -1,0 +1,14 @@
+package automation;
+
+import org.junit.platform.suite.api.ConfigurationParameter;
+import org.junit.platform.suite.api.IncludeEngines;
+import org.junit.platform.suite.api.SelectClasspathResource;
+import org.junit.platform.suite.api.Suite;
+
+import static io.cucumber.core.options.Constants.GLUE_PROPERTY_NAME;
+
+@Suite
+@IncludeEngines("cucumber")
+@SelectClasspathResource("automation/steps")
+@ConfigurationParameter(key = GLUE_PROPERTY_NAME, value = "automation.steps")
+public class CucumberTest {}
